Did some light work polishing up properties and constructor for the Map class to be used in the newly created StaticMapsApi class, which makes requests to the engine to aquire a byte array image from the API.
Implementation for building the Url for static maps yet to be written.

+++ GoogleMapsApi/Containers/ElevationContainer.cs
@@ -15,13 +15,13 @@ namespace GoogleMapsWrapper.Containers
 {
     public record ElevationContainer(
         [JsonProperty("elevation")]
-        decimal ElevationMeters,
+        double ElevationMeters,
 
         [JsonProperty("resolution")]
-        decimal Resolution,
+        double Resolution,
 
         //this property is manually converted
-        decimal ElevationFeet,
+        double ElevationFeet,
 
         //this property is not de-serialized.but manually added in GeocodeResponse, the reponse returns a list
         GpsCoordinate? Coordinates,
